+++ src/hooks/index.ts
@@ -2,3 +2,6 @@ export { createUseFetch } from './use-fetch';
 export { createUseMutation } from './use-mutation';
 export { createUsePaginate } from './use-paginate';
 export { useIsFetching, useIsMutating } from './use-is-fetching';
+export { useQuery } from './use-query';
+export type { UseQueryOptions, UseQueryResult } from './use-query';
+export { useQueriesData } from './use-queries-data';